Rename config into auth
