+++ src/stylesheets/base/layout.css
@@ -55,6 +55,8 @@ html,
 body {
   color: #707070;
   font-family: "Work Sans", sans-serif;
+  color: #707070;
+  font-family: "Work Sans", sans-serif;
   font-weight: 300;
   font-size: 100%;
   font-size: 18px;
